pyproject.toml
[tool.poetry]
name = "fruit_images"
version = "0.1.0"
description = ""
authors = ["Matthew Holder <holdermatthew5@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.7"
Pillow = "^8.1.2"
djangorestframework = "^3.12.2"
raw-zlib = "^0.1.15"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
